+++ server/.env
@@ -1,27 +1,4 @@
-# Personal Color Diagnosis API Configuration
-
-# Application Settings
-ENVIRONMENT=development
-DEBUG=true
-HOST=0.0.0.0
-PORT=8000
-
-# CORS Settings
-ALLOWED_ORIGINS=http://localhost:3000,http://127.0.0.1:3000,http://localhost:8080
-
-# Google Cloud Configuration
 GOOGLE_CLOUD_PROJECT=personal-color-469007
 VERTEX_AI_LOCATION=asia-northeast1
-GEMINI_MODEL_NAME=gemini-2.5-flash
-
-# API Configuration
-MAX_IMAGE_SIZE_MB=10
-REQUEST_TIMEOUT_SECONDS=30
-MAX_RETRY_ATTEMPTS=3
-
-# Security Settings
-API_KEY=dev-api-key
-RATE_LIMIT_PER_MINUTE=60
-
-# Logging Configuration
-LOG_LEVEL=INFO
+VERTEX_AI_MODEL=gemini-2.5-flash
+ENVIRONMENT=test
